--- a/be/src/common/config.h
+++ b/be/src/common/config.h
@@ -1582,4 +1582,9 @@ CONF_mInt32(put_combined_txn_log_thread_pool_num_max, "64");
 CONF_mBool(enable_put_combinded_txn_log_parallel, "false");
 // used to control whether the metrics/ interface collects table metrics
 CONF_mBool(enable_collect_table_metrics, "true");
+// some internal parameters are used to control the execution strategy of join runtime filter pushdown.
+// Do not modify them unless necessary.
+CONF_mInt64(rf_sample_rows, "1024");
+CONF_mInt64(rf_sample_ratio, "32");
+CONF_mInt64(rf_branchless_ratio, "8");
 } // namespace starrocks::config