--- a/be/src/exprs/runtime_filter_layout.h
+++ b/be/src/exprs/runtime_filter_layout.h
@@ -127,4 +127,31 @@ auto dispatch_layout(bool global, const RuntimeFilterLayout& layout, Args&&... a
     }
 }
 
+template <template <TRuntimeFilterLayoutMode::type> typename F, typename... Args>
+auto dispatch_layout_v2(bool global, const RuntimeFilterLayout& layout, Args&&... args) {
+    TRuntimeFilterLayoutMode::type layout_mode = global ? layout.global_layout() : layout.local_layout();
+    switch (layout_mode) {
+    case TRuntimeFilterLayoutMode::SINGLETON:
+        return F<TRuntimeFilterLayoutMode::SINGLETON>()(layout, std::forward<Args>(args)...);
+    case TRuntimeFilterLayoutMode::PIPELINE_SHUFFLE:
+        return F<TRuntimeFilterLayoutMode::PIPELINE_SHUFFLE>()(layout, std::forward<Args>(args)...);
+    case TRuntimeFilterLayoutMode::GLOBAL_SHUFFLE_2L:
+        return F<TRuntimeFilterLayoutMode::GLOBAL_SHUFFLE_2L>()(layout, std::forward<Args>(args)...);
+    case TRuntimeFilterLayoutMode::GLOBAL_SHUFFLE_1L:
+        return F<TRuntimeFilterLayoutMode::GLOBAL_SHUFFLE_1L>()(layout, std::forward<Args>(args)...);
+    case TRuntimeFilterLayoutMode::PIPELINE_BUCKET_LX:
+        return F<TRuntimeFilterLayoutMode::PIPELINE_BUCKET_LX>()(layout, std::forward<Args>(args)...);
+    case TRuntimeFilterLayoutMode::PIPELINE_BUCKET:
+        return F<TRuntimeFilterLayoutMode::PIPELINE_BUCKET>()(layout, std::forward<Args>(args)...);
+    case TRuntimeFilterLayoutMode::GLOBAL_BUCKET_2L_LX:
+        return F<TRuntimeFilterLayoutMode::GLOBAL_BUCKET_2L_LX>()(layout, std::forward<Args>(args)...);
+    case TRuntimeFilterLayoutMode::GLOBAL_BUCKET_2L:
+        return F<TRuntimeFilterLayoutMode::GLOBAL_BUCKET_2L>()(layout, std::forward<Args>(args)...);
+    case TRuntimeFilterLayoutMode::GLOBAL_BUCKET_1L:
+        return F<TRuntimeFilterLayoutMode::GLOBAL_BUCKET_1L>()(layout, std::forward<Args>(args)...);
+    default:
+        CHECK(false) << "Unknown type: " << layout_mode;
+        __builtin_unreachable();
+    }
+}
 } // namespace starrocks